added master checkbox with confirmation dialog

+++ app/javascript/application.js
@@ -28,11 +28,29 @@ $(document).ready(function() {
       }
     });
 
-    $.ajax({
-      url: 'books/bulk_delete_books',
-      type: 'DELETE',
-      data: { book_ids: bookIds }
-    });
+    if(confirm('Are you sure you want to delete selected books? This action can not be undone. If you want to delete selected books click on the OK else click on Cancel.')) {
+      $.ajax({
+        url: 'books/bulk_delete_books',
+        type: 'DELETE',
+        data: { book_ids: bookIds }
+      });
+    } else {
+      return false;
+    }
+  });
+
+  $("#books-index-master").on('click', function(){
+    $('.book-select-check').prop('checked', $(this).prop('checked'))
+
+    var counter = 0;
+
+    if($(this).prop('checked')) {
+      counter += $('.book-select-check').length;
+      $("#delete-books").text(`Delete ${pluralize(counter, 'Book')}`);
+      $("#delete-books").show()
+    } else {
+      $("#delete-books").hide()
+    }
   });
 
 });
